HAL/requirements.txt
numpy
Pillow
networkx
nltk
opencv-python
reportlab
ipywidgets
matplotlib
